tools/spelling-checker/requirements.txt
PyPDF2==3.0.1
pyspellchecker==0.8.1
language-tool-python==2.8.1
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.3.0
